requirements.txt
click
passlib
prettytable
psycopg2-binary
dbus-python
sphinx-click
cookiecutter
pycodestyle
